requirements-macos.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
confluent-kafka==2.10.0
docker==7.1.0
etcd3==0.12.0
Flask==3.1.1
grpcio==1.71.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
protobuf==3.20.3
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.1
six==1.17.0
tenacity==9.1.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
